fix: resolve all ares_code_lint findings (1 ERROR + 39 WARNINGs) — v2…

+++ src/ams/mission_script_engine_parser_meta.cpp
@@ -486,8 +486,8 @@ bool MissionScriptEngine::parsePulseMinAltLineLocked(const char* line)
     if (*rest == '\0') { setErrorLocked("pulse.min_altitude: missing value"); return false; }
 
     char* endp = nullptr;
-    const long val = strtol(rest, &endp, 10);
-    if (endp == rest || val <= 0L || val > 50000L)
+    const int64_t val = static_cast<int64_t>(strtol(rest, &endp, 10));
+    if (endp == rest || val <= 0 || val > 50000)
     {
         setErrorLocked("pulse.min_altitude: value must be in range [1, 50000] m");
         return false;
@@ -526,8 +526,8 @@ bool MissionScriptEngine::parsePulseSafeDelayLineLocked(const char* line)
     if (*rest == '\0') { setErrorLocked("pulse.safe_delay: missing value"); return false; }
 
     char* endp = nullptr;
-    const long val = strtol(rest, &endp, 10);
-    if (endp == rest || val < 100L || val > 60000L)
+    const int64_t val = static_cast<int64_t>(strtol(rest, &endp, 10));
+    if (endp == rest || val < 100 || val > 60000)
     {
         setErrorLocked("pulse.safe_delay: value must be in range [100, 60000] ms");
         return false;
@@ -565,8 +565,8 @@ bool MissionScriptEngine::parsePulseArmTimeoutLineLocked(const char* line)
     if (*rest == '\0') { setErrorLocked("pulse.arm_timeout: missing value"); return false; }
 
     char* endp = nullptr;
-    const long val = strtol(rest, &endp, 10);
-    if (endp == rest || val < 500L || val > 300000L)
+    const int64_t val = static_cast<int64_t>(strtol(rest, &endp, 10));
+    if (endp == rest || val < 500 || val > 300000)
     {
         setErrorLocked("pulse.arm_timeout: value must be in range [500, 300000] ms");
         return false;
